services/analyzer/pyproject.toml
[project]
name = "analyzer"
version = "0.0.1"
description = "Analyzer Microservice"
readme = "README.md"
requires-python = ">=3.13, <3.14"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "instructor>=1.11.3",
    "keybert>=0.8.5",
    "mlflow==3.10.1",
    "openai>=1.108.1",
    "opentelemetry-exporter-otlp-proto-http>=1.34.1",
    "opentelemetry-exporter-prometheus>=0.61b0",
    "opentelemetry-instrumentation-fastapi>=0.55b1",
    "opentelemetry-sdk>=1.34.1",
    "prometheus-client>=0.24.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "sentence-transformers>=3.0.0",
    "sqlalchemy>=2.0.41",
]

[dependency-groups]
dev = [
    "pact-python>=3.2.1",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "testcontainers[postgres]>=4.10.0",
]
local = ["ipykernel>=6.29.5", "ruff>=0.12.4", "ty>=0.0.1a15"]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
    'ignore:function ham\(\) is deprecated:DeprecationWarning',
    # OTel BatchSpanProcessor background thread can raise ValueError when
    # trying to log after pytest closes its streams. Suppress this so it
    # doesn't inflate the exit code via pytest's threading.excepthook capture.
    "ignore::pytest.PytestUnraisableExceptionWarning",
]
testpaths = ["tests"]
markers = [
    "unit: fast unit tests with mocks/direct function calls",
    "integration: tests that exercise HTTP handlers, database integration, or app wiring",
    "contract: Pact/provider contract verification tests",
    "wip: work in progress",
    "infrastructure: mocks for dummy data (sql db, http endpoint etc)",
]
addopts = "--cov=src --cov-report=term-missing --color=yes"
